Refactor rest of IO
