--- syftlib/src/test/java/org/openmined/syft/unit/utilities/IOTest.kt
@@ -1,4 +1,0 @@
-package org.openmined.syft.unit.utilities
-
-internal class IOTest {
-}
